Fix remark about pushing callee-saved registers
This file had the following remark:
"Because even if this function doesn't change a certain callee-saved
register, the function we'll call may change it, so we must push them
all."
This is incorrect: a non-leaf function doesn't have to push any
registers that the function itself won't change. If one of the callers
change that register, they must revert it before returning.
Remove that remark and expand on the possibility of using the stack
versus registers for local variables.